requirements.txt
django
psycopg2-binary
pillow
djangorestframework
python-dotenv
django-filter